move toolbar into its own component
